rewrite vue app to typescript
