+++ js/package.json
@@ -2,24 +2,34 @@
   "name": "supercollider_web_rtc_gui",
   "version": "0.1.0",
   "private": true,
+  "description": "Allows to dynamically create GUIs for the web via SuperCollider",
+  "author": "Dennis Scheiba",
   "scripts": {
     "serve": "vue-cli-service serve",
     "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint"
+    "lint": "vue-cli-service lint",
+    "client": "ts-node --esm src/client.ts",
+    "server": "ts-node --esm src/server.ts",
+    "local": "npm run client & npm run server & npm run serve"
   },
   "dependencies": {
     "@mdi/font": "5.9.55",
     "core-js": "^3.8.3",
+    "express": "^4.17.3",
+    "osc": "^2.4.2",
     "roboto-fontface": "*",
+    "socket.io": "^4.4.1",
     "socket.io-client": "^4.4.1",
     "vue": "^3.2.13",
     "vue-class-component": "^8.0.0-0",
-    "vuetify": "^3.0.0-beta.0",
+    "vuetify": "^3.0.4",
     "webfontloader": "^1.0.0"
   },
   "devDependencies": {
     "@babel/core": "^7.12.16",
     "@babel/eslint-parser": "^7.12.16",
+    "@types/express": "^4.17.14",
+    "@types/node": "^18.11.9",
     "@types/webfontloader": "^1.6.35",
     "@typescript-eslint/eslint-plugin": "^5.4.0",
     "@typescript-eslint/parser": "^5.4.0",
@@ -32,8 +42,18 @@
     "eslint-plugin-vue": "^8.0.3",
     "sass": "^1.38.0",
     "sass-loader": "^10.0.0",
+    "ts-node": "^10.9.1",
     "typescript": "~4.5.5",
     "vue-cli-plugin-vuetify": "~2.4.8",
     "vuetify-loader": "^2.0.0-alpha.0"
+  },
+  "bugs": {
+    "url": "https://github.com/capital-G/webRTCgui/issues"
+  },
+  "homepage": "https://github.com/capital-G/webRTCgui#readme",
+  "license": "GPL-2.0",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/capital-G/webRTCgui.git"
   }
 }
