+++ js/src/components/ControllerButton.vue
@@ -1,17 +1,17 @@
 <script setup lang="ts">
 import { defineProps, ref, Ref, PropType } from "vue";
 import { useSocketIO } from "../services/socketio.service";
-import { TextController } from "../../../communication";
+import { ButtonController } from "../communication";
 
 const { socket } = useSocketIO();
 
 const props = defineProps({
-  controller: {type: Object as PropType<TextController>, required: true}
+  controller: {type: Object as PropType<ButtonController>, required: true}
 });
 
-var value: Ref<string> = ref(props.controller.value);
+var value: Ref<number> = ref(props.controller.value);
 
-async function updateText() {
+async function buttonPress() {
     // as we can not update props in vue
     // we instead copy it and use it to set the values
     var c = structuredClone({...props.controller});
@@ -25,13 +25,11 @@ async function updateText() {
     <v-row class="text-center">
       <v-col>
         <div>
-          <div class="text-caption">{{ $props.controller.name }}</div>
-          <v-textarea
+          <v-btn
             v-model="value"
-            @change="updateText()"
+            @click="buttonPress()"
             color="orange"
-            label="color"
-          />
+          >{{ $props.controller.name }}</v-btn>
         </div>
       </v-col>
     </v-row>
