+++ js/src/main.ts
@@ -1,0 +1,18 @@
+import { createApp } from 'vue'
+import App from './App.vue'
+import { loadFonts } from './plugins/webfontloader'
+
+// Vuetify
+import 'vuetify/styles'
+import { createVuetify } from 'vuetify'
+import * as components from 'vuetify/components'
+import * as directives from 'vuetify/directives'
+
+loadFonts()
+
+const vuetify = createVuetify({
+  components,
+  directives,
+})
+
+createApp(App).use(vuetify).mount('#app')
