+++ js/src/shims-vuetify.d.ts
@@ -1,0 +1,3 @@
+declare module 'vuetify'
+declare module 'vuetify/lib/components'
+declare module 'vuetify/lib/directives'
