--- package.json
@@ -1,31 +1,0 @@
-{
-  "name": "web_rtc_gui",
-  "version": "0.1.0",
-  "description": "Allows to dynamically create GUIs for the web via SuperCollider",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "server": "ts-node --esm server.ts",
-    "client": "ts-node --esm client.ts"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/capital-G/webRTCgui.git"
-  },
-  "author": "Dennis Scheiba",
-  "license": "GPL-2.0",
-  "bugs": {
-    "url": "https://github.com/capital-G/webRTCgui/issues"
-  },
-  "homepage": "https://github.com/capital-G/webRTCgui#readme",
-  "dependencies": {
-    "express": "^4.17.3",
-    "osc": "^2.4.2",
-    "socket.io": "^4.4.1",
-    "socket.io-client": "^4.4.1"
-  },
-  "devDependencies": {
-    "@types/express": "^4.17.14",
-    "@types/node": "^18.11.9",
-    "ts-node": "^10.9.1"
-  }
-}
